添加nacos,feign,通用日志切面等

+++ common/src/main/java/cn/datawisher/common/service/CurdService.java
@@ -40,7 +40,7 @@ public interface CurdService<T> {
      * @param id
      * @return
      */
-    T findById(Long id);
+    T findById(Integer id);
 
     /**
      * 分页查询
